+++ Civi/ActionProvider/Provider.php
@@ -70,6 +70,7 @@ class Provider {
     Action\MailingEvent\Actions::loadActions($this);
     Action\Communication\Actions::loadActions($this);
     Action\Generic\Actions::loadActions($this);
+    Action\Tag\Actions::loadActions($this);
 
 		$conditions = array(
 		  new \Civi\ActionProvider\Condition\ParameterIsEmpty(),